# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/pages/{slug}": {
      "get": {
        "operationId": "pagesBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicPage"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "The public page at vrc.page/<slug>. Private, hidden, held and unknown names are one identical 404.",
        "tags": [
          "pages"
        ]
      }
    },
    "/v1/me/dashboard": {
      "get": {
        "operationId": "meDashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "The dashboard's frame: the account's page, its groups, and what is waiting for it.",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/page": {
      "get": {
        "operationId": "mePage",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnUserPage"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "The account's own page, whatever its visibility; 404 before VRChat is connected.",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/groups/{pageId}": {
      "get": {
        "operationId": "meGroup",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnGroupPage"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "A group page this account owns or edits; 404 for any other.",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/notification-preferences": {
      "get": {
        "operationId": "meNotificationPreferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferences"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Which emails the account gets.",
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "meSetNotificationPreferences",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationPreferencesPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferences"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Change some of the notification preferences; the answer is all of them.",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/pages/{pageId}/visibility": {
      "put": {
        "operationId": "meSetVisibility",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisibilityRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Who can open one of this account's pages. Owners only: an editor could\notherwise unpublish a community's page. A page this account doesn't run\nis the same 404 as one that isn't there.",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/vrchat": {
      "delete": {
        "operationId": "meDisconnectVRChat",
        "parameters": [],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Disconnect VRChat. The page goes, and so do the groups claimed through\nthat connection and their pages; every name is held for 90 days.",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/names/{name}": {
      "get": {
        "operationId": "meNameAvailability",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NameAvailability"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Whether a name can be used. Signed in, because a stranger has no business\nmapping which names exist. `pageId` names the page asking, so its own name\ncomes back as \"yours\" rather than \"taken\".",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/pages/{pageId}/name": {
      "put": {
        "operationId": "meSetName",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageName"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Give one of this account's pages its name, or change it. Owners only. The\nfirst name is free; after a change the name settles for slug.change_cooldown_days,\nand the one it replaced is held so nobody can pick it up to pass as its owner.",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/pages/{pageId}/refresh": {
      "post": {
        "operationId": "meRefresh",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Refreshed"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Read the page again from VRChat now, rather than waiting for its turn.\nOwners only, one manual refresh per page per refresh.manual.cooldown_seconds,\nand refresh.manual.daily_cap_per_account a day.",
        "tags": [
          "me"
        ]
      }
    },
...
  },
  "info": {
    "title": "vrc.page API",
    "description": "The backend for vrc.page. Errors use RFC 9457 Problem Details (application/problem+json), and every response carries an X-Request-Id header.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UserPage": {
        "type": "object",
        "properties": {
          "pronouns": {
            "type": "string",
            "nullable": true
          },
          "trustRank": {
            "type": "string",
            "nullable": true
          },
          "group": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/RepresentedGroup"
              }
            ]
          },
          "status": {
            "enum": [
              "active",
              "join_me",
              "ask_me",
              "busy",
              "offline"
            ],
            "type": "string"
          },
          "statusDescription": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true,
            "description": "Plain text with its line breaks."
          },
          "bannerUrl": {
            "type": "string",
            "nullable": true,
            "description": "Our own copy, never a VRChat address. Null until images are stored."
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "visibility": {
            "enum": [
              "public",
              "unlisted",
              "private"
            ],
            "type": "string"
          },
          "vrchatUserId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "ageVerified": {
            "type": "boolean"
          },
          "links": {
            "description": "VRChat's first, then the ones added on vrc.page, in their order.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PageLink"
            }
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "verifiedAt": {
            "type": "string",
            "description": "When the bio code matched: ISO 8601."
          },
          "lastRefreshedAt": {
            "type": "string",
            "description": "Last successful read from VRChat: ISO 8601."
          }
        },
        "required": [
          "pronouns",
          "trustRank",
          "group",
          "status",
          "statusDescription",
          "bio",
          "bannerUrl",
          "avatarUrl",
          "visibility",
          "vrchatUserId",
          "displayName",
          "ageVerified",
          "links",
          "languages",
          "verifiedAt",
          "lastRefreshedAt"
        ]
      },
...
      "GroupPage": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "rules": {
            "type": "string",
            "nullable": true
          },
          "memberCount": {
            "type": "integer"
          },
          "iconUrl": {
            "type": "string",
            "nullable": true
          },
          "bannerUrl": {
            "type": "string",
            "nullable": true
          },
          "owner": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/GroupOwner"
              }
            ]
          },
          "visibility": {
            "enum": [
              "public",
              "unlisted",
              "private"
            ],
            "type": "string"
          },
          "vrchatGroupId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "shortCode": {
            "type": "string",
            "description": "Shown with the discriminator as ABCDE.1234, the code VRChat uses."
          },
          "discriminator": {
            "type": "string"
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PageLink"
            }
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isVerified": {
            "type": "boolean",
            "description": "VRChat's own verification."
          },
          "verifiedAt": {
            "type": "string",
            "description": "When the claim succeeded: ISO 8601."
          },
          "lastRefreshedAt": {
            "type": "string"
          }
        },
        "required": [
          "description",
          "rules",
          "memberCount",
          "iconUrl",
          "bannerUrl",
          "owner",
          "visibility",
          "vrchatGroupId",
          "name",
          "shortCode",
          "discriminator",
          "links",
          "languages",
          "isVerified",
          "verifiedAt",
          "lastRefreshedAt"
        ]
      },
      "PublicPage": {
        "type": "object",
        "properties": {
          "kind": {
            "enum": [
              "user",
              "group"
            ],
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserPage"
          },
          "group": {
            "$ref": "#/components/schemas/GroupPage"
          },
          "slug": {
            "type": "string",
            "description": "The page's own name, as its owner typed it."
          },
          "alias": {
            "type": "boolean",
            "description": "The name asked for is an alias: redirect (308) to `slug`."
          }
        },
        "required": [
          "kind",
          "slug",
          "alias"
        ]
      },
      "DashboardPage": {
        "type": "object",
        "properties": {
          "kind": {
            "enum": [
              "user",
              "group"
            ],
            "type": "string"
          },
          "slug": {
            "type": "string",
            "nullable": true,
            "description": "Null until a name is chosen."
          },
          "visibility": {
            "enum": [
              "public",
              "unlisted",
              "private"
            ],
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "editor"
            ],
            "type": "string"
          },
          "iconUrl": {
            "type": "string",
            "nullable": true
          },
          "memberCount": {
            "type": "integer",
            "nullable": true,
            "description": "Groups only."
          },
          "id": {
            "type": "string",
            "description": "vrc.page's page id, used in dashboard addresses."
          },
          "name": {
            "type": "string"
          },
          "vrchatId": {
            "type": "string",
            "description": "The usr_ or grp_ id, which seeds the fallback colour."
          }
        },
        "required": [
          "kind",
          "slug",
          "visibility",
          "role",
          "iconUrl",
          "memberCount",
          "id",
          "name",
          "vrchatId"
        ]
      },
      "Dashboard": {
        "type": "object",
        "properties": {
          "user": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DashboardPage"
              }
            ]
          },
          "groupClaimsLeft": {
            "type": "integer",
            "description": "More groups this account may still claim."
          },
          "pendingInvites": {
            "type": "integer"
          },
          "groups": {
            "description": "Owned groups first, then the ones this account edits.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DashboardPage"
            }
          }
        },
        "required": [
          "user",
          "groupClaimsLeft",
          "pendingInvites",
          "groups"
        ]
      },
      "OwnUserPage": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "nullable": true
          },
          "nameChangeableAt": {
            "type": "string",
            "nullable": true,
            "description": "When the page's name may next change: ISO 8601, or null when it can now."
          },
          "refreshableAt": {
            "type": "string",
            "nullable": true,
            "description": "When it may next be refreshed from VRChat by hand: ISO 8601, or null when it can now."
          },
          "pageId": {
            "type": "string"
          },
          "page": {
            "$ref": "#/components/schemas/UserPage"
          }
        },
        "required": [
          "slug",
          "nameChangeableAt",
          "refreshableAt",
          "pageId",
          "page"
        ]
      },
      "OwnGroupPage": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "enum": [
              "owner",
              "editor"
            ],
            "type": "string"
          },
          "nameChangeableAt": {
            "type": "string",
            "nullable": true,
            "description": "When the page's name may next change: ISO 8601, or null when it can now."
          },
          "refreshableAt": {
            "type": "string",
            "nullable": true,
            "description": "When it may next be refreshed from VRChat by hand: ISO 8601, or null when it can now."
          },
          "pageId": {
            "type": "string"
          },
          "page": {
            "$ref": "#/components/schemas/GroupPage"
          }
        },
        "required": [
          "slug",
          "role",
          "nameChangeableAt",
          "refreshableAt",
          "pageId",
          "page"
        ]
      },
      "NotificationPreferences": {
        "type": "object",
        "properties": {
          "groupInvites": {
            "type": "boolean",
            "description": "Someone invited you to edit a group."
          },
          "pageChanges": {
            "type": "boolean",
            "description": "Your page's name, visibility or links changed."
          },
          "productNews": {
            "type": "boolean",
            "description": "Occasional news about vrc.page itself."
          }
        },
        "required": [
          "groupInvites",
          "pageChanges",
          "productNews"
        ]
      },
      "VisibilityRequest": {
        "type": "object",
        "properties": {
          "visibility": {
            "enum": [
              "public",
              "unlisted",
              "private"
            ],
            "type": "string"
          }
        },
        "required": [
          "visibility"
        ]
      },
      "NotificationPreferencesPatch": {
        "type": "object",
        "properties": {
          "groupInvites": {
            "type": "boolean"
          },
          "pageChanges": {
            "type": "boolean"
          },
          "productNews": {
            "type": "boolean"
          }
        }
      },
      "NameAvailability": {
        "type": "object",
        "properties": {
          "status": {
            "enum": [
              "available",
              "yours",
              "taken",
              "held",
              "reserved",
              "impersonation",
              "invalid",
              "too_short",
              "too_long"
            ],
            "type": "string",
            "description": "available: free. yours: this page already has it. taken: another page has it. held: released recently and still held. reserved: a name the site keeps. impersonation: contains a blocked word. invalid: not letters, digits, _ or -."
          },
          "minLength": {
            "type": "integer"
          },
          "maxLength": {
            "type": "integer"
          }
        },
        "required": [
          "status",
          "minLength",
          "maxLength"
        ]
      },
      "NameRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "PageName": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "slug"
        ]
      },
      "Refreshed": {
        "type": "object",
        "properties": {
          "refreshedAt": {
            "type": "string",
            "description": "ISO 8601."
          }
        },
        "required": [
          "refreshedAt"
        ]
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "The HTTP status code.",
            "example": 404
          },
          "retryAfter": {
            "type": "integer",
            "description": "Seconds to wait before trying again, for a cooldown."
          },
          "at": {
            "type": "integer",
            "description": "Which item of a submitted list the problem is about, counting from 0."
          },
          "type": {
            "type": "string",
            "description": "A URI naming the kind of problem: \"about:blank\" when the status says it\nall, otherwise https://vrc.page/problems/<code>, with a code from\nsrc/common/problem.ts."
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary of the kind of problem."
          },
          "detail": {
            "type": "string",
            "description": "What went wrong this time, in words a person can act on."
          },
          "instance": {
            "type": "string",
            "description": "The path of the request that failed."
          },
          "requestId": {
            "type": "string",
            "description": "Same as the X-Request-Id response header and the request's audit log entries."
          }
        },
        "required": [
          "status",
          "type",
          "title",
          "requestId"
        ]
      }
    }
  }
}